refactor: Consolidate workspace members
