+++ rust/operator-binary/src/discovery.rs
