+++ rust/operator-binary/src/event.rs
