+++ rust/operator-binary/src/operations/graceful_shutdown.rs
