+++ rust/operator-binary/src/operations/pdb.rs
