+++ rust/operator-binary/src/product_logging.rs
