--- rust/operator/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "stackable-hdfs-operator"
-description = "Stackable Operator for Apache Hadoop HDFS - The Operator Library"
-version.workspace = true
-authors.workspace = true
-license.workspace = true
-edition.workspace = true
-repository.workspace = true
-publish = false
-
-[dependencies]
-stackable-hdfs-crd = { path = "../crd" }
-
-futures.workspace = true
-indoc.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-snafu.workspace = true
-stackable-operator.workspace = true
-product-config.workspace = true
-strum.workspace = true
-tokio.workspace = true
-tracing.workspace = true
-tracing-futures.workspace = true
-
-[dev-dependencies]
-rstest.workspace = true
